fix(ops): move restoreHiddenFactRows to facts-fence.ts (module-size ratchet)
CI's check:module-size job failed on this branch's own diff:
src/core/import-file.ts grew to 2178 lines, 36 over its 2142 committed
ceiling (scripts/module-size-limits.tsv) after the previous commit added
restoreHiddenFactRows() there.

Per the guard's own preferred remedy ("move the new code into a sibling
module"), restoreHiddenFactRows() moves to facts-fence.ts, next to the
ParsedFact/parseFactsFence/renderFactsTable primitives it composes and
already depends on. import-file.ts now imports it instead of defining it,
and the two call-site comments are condensed (the detailed P1/P2 rationale
now lives once, in the function's own docstring, rather than duplicated at
each call site).

No behavior change. Local verification: bun run verify (54/54 checks,
including check:module-size) green; typecheck clean; the full
privacy-strip-and-forget + import-file + facts-fence* + extract-facts-phase
test files (143 tests total) pass unchanged.
